pyproject.toml
[build-system]
requires = ['setuptools', 'wheel']
build-backend = "setuptools.build_meta"

[project]
name = "lastfm-oss-analytics"
version = "0.1.0"
dependencies = [
"requests",
"pydantic",
    "pydantic-settings",
    "apache-airflow",
"pandas",
"sqlalchemy",
"psycopg[binary]",
"boto3",
"python-dotenv",
"rich",
]

[project.optional-dependencies]
dev = ["pytest", "pre-commit"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"

[tool.black]
line-length=79
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 79
# optional niceties:
combine_as_imports = true
force_sort_within_sections = true
